server/requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.3
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
decorator==4.4.2
Django==3.2.4
django-filter==21.1
django-mongoengine==0.4.6
django-rest-knox==4.1.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
excel2json-3==0.1.6
Flask==2.0.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.11.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==2.0.1
mongoengine==0.23.1
mysqlclient @ file:///D:/%D0%A3%D1%87%D0%B5%D0%B1%D0%B0/%D0%92%D0%A3%D0%97/4%20%D0%BA%D1%83%D1%80%D1%81/8%20%D1%81%D0%B5%D0%BC%D0%B5%D1%81%D1%82%D1%80/XP/project/server/mysqlclient-1.4.6-cp38-cp38-win32.whl
openpyxl==3.0.9
parsel==1.6.0
Pillow==9.1.0
priority==1.3.0
Protego==0.1.16
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.4
pyOpenSSL==21.0.0
python-telegram-bot==13.10
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.26.0
rsa==4.8
Scrapy==2.5.0
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
telegram==0.0.1
Telethon==1.24.0
tornado==6.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vk==2.0.2
w3lib==1.22.0
Werkzeug==2.0.1
xlrd==2.0.1
zipp==3.7.0
zope.interface==5.4.0
